+++ example/example.dart
@@ -3,21 +3,12 @@ import 'package:simdart/simdart.dart';
 void main() async {
   final SimDart sim = SimDart(onTrack: (track) => print(track));
 
-  sim.process(eventA, name: 'A');
-  sim.process(eventB, start: 5, name: 'B');
+  sim.process(event: _a, name: 'A');
 
-  await sim.run();
+  await sim.run(until: 10);
 }
 
-void eventA(EventContext context) async {
-  await context.wait(10);
-  context.sim.process(eventC, delay: 1, name: 'C');
-}
-
-void eventB(EventContext context) async {
-  await context.wait(1);
-}
-
-void eventC(EventContext context) async {
-  await context.wait(10);
+void _a(EventContext context) async {
+  await context.wait(2);
+  context.sim.process(event: _a, delay: 2, name: 'A');
 }